[config] post 생성 초기 설정 완료
