+++ posts/mdtest.mdx
@@ -1,0 +1,19 @@
+---
+title: 길종님은 짱짱맨이신가요?
+description: 네 그렇다고하네요
+categories: 개발
+writeDate: 2024-05-16T00:00:00Z
+releaseDate: 2024-05-17T00:00:00Z
+authorName: 김명진
+authorIcon: /authorIcon/testIcon.png
+authorIntro: 김명진입니다......
+---
+
+# 이해하였습니다.
+## 길종님 짱짱맨
+- li
+- li
+
+```
+| 우왕!
+
